+++ users-service/src/main/kotlin/piperkt/services/users/Application.kt
@@ -1,0 +1,7 @@
+package piperkt.services.users
+
+import io.micronaut.runtime.Micronaut.run
+
+fun main(vararg args: String) {
+    run(*args)
+}