+++ users-service/src/main/kotlin/piperkt/services/users/infrastructure/Utils.kt
@@ -1,0 +1,7 @@
+package piperkt.services.users.infrastructure
+
+import java.util.*
+
+object Utils {
+    fun <T> Optional<T>.asNullable(): T = orElse(null)
+}