+++ users-service/src/main/kotlin/piperkt/services/users/infrastructure/implementation/Services.kt
@@ -1,0 +1,18 @@
+package piperkt.services.users.infrastructure.implementation
+
+import jakarta.inject.Singleton
+import piperkt.services.users.application.AuthService
+import piperkt.services.users.application.UserService
+import piperkt.services.users.domain.user.UserEventPublisher
+import piperkt.services.users.domain.user.UserRepository
+
+/** The services implementations powered by Micronaut DI. */
+object Services {
+    @Singleton
+    class UserServiceImpl(userRepository: UserRepository, userEventPublisher: UserEventPublisher) :
+        UserService(userRepository, userEventPublisher)
+
+    @Singleton
+    class AuthServiceImpl(userRepository: UserRepository, userEventPublisher: UserEventPublisher) :
+        AuthService(userRepository, userEventPublisher)
+}