+++ users-service/src/main/kotlin/piperkt/services/users/infrastructure/persistence/repository/UserRepositoryImpl.kt
@@ -1,0 +1,42 @@
+package piperkt.services.users.infrastructure.persistence.repository
+
+import jakarta.inject.Singleton
+import piperkt.services.users.domain.user.User
+import piperkt.services.users.domain.user.UserRepository
+import piperkt.services.users.domain.user.Username
+import piperkt.services.users.infrastructure.persistence.model.UserEntityMapper.toDomain
+import piperkt.services.users.infrastructure.persistence.model.UserEntityMapper.toEntity
+import piperkt.services.users.infrastructure.persistence.model.UserEntityRepository
+
+@Singleton
+class UserRepositoryImpl(private val userEntityRepository: UserEntityRepository) : UserRepository {
+    override fun findByUsername(username: String): User? {
+        return userEntityRepository.findByUsername(username)?.toDomain()
+    }
+
+    override fun findByRefreshToken(refreshToken: String): User? {
+        return userEntityRepository.findByRefreshToken(refreshToken)?.toDomain()
+    }
+
+    override fun findById(id: Username): User? {
+        return userEntityRepository.findByUsername(id.value)?.toDomain()
+    }
+
+    override fun save(entity: User) {
+        userEntityRepository.save(entity.toEntity())
+    }
+
+    override fun deleteById(id: Username): User? {
+        val user = findById(id)
+        userEntityRepository.deleteByUsername(id.value)
+        return user
+    }
+
+    override fun update(entity: User) {
+        userEntityRepository.save(entity.toEntity())
+    }
+
+    override fun deleteAll() {
+        userEntityRepository.deleteAll()
+    }
+}