+++ users-service/src/test/kotlin/architecture/Architecture.kt
@@ -1,0 +1,7 @@
+package architecture
+
+const val PREFIX = "piperkt.services.users"
+
+class CleanArchitectureMultimediaTest : CleanArchitectureSpec(PREFIX)
+
+class FrameworkIndependenceMultimediaTest : FrameworkIndependenceTest(PREFIX)